requirements.txt
python-fitparse
